remove ProvisioningParams from model
